Bug fixed where over 256 bytes of input crashed serial driver.

+++ back/serial.c
@@ -1,5 +1,7 @@
 #include "serial.h"
 
+#include <stdio.h>              // XXX
+
 #include <util/atomic.h>
 
 #include "e-stop.h"
@@ -18,17 +20,17 @@
 #define ASCII_XON  '\021'
 #define ASCII_XOFF '\023'
 
-static uint8_t  tx_buf[TX_BUF_SIZE] __attribute__((aligned(256)));
-static uint8_t  tx_head;
-static uint8_t  tx_tail;
-static uint8_t  tx_errs;
-static uint8_t  tx_oob_char;
+static uint8_t tx_buf[TX_BUF_SIZE] __attribute__((aligned(256)));
+static uint8_t tx_head;
+static uint8_t tx_tail;
+static uint8_t tx_errs;
+static uint8_t tx_oob_char;
 
-static uint8_t  rx_buf[RX_BUF_SIZE] __attribute__((aligned(256)));
-static uint16_t rx_head;
-static uint16_t rx_tail;
-static uint8_t  rx_errs;
-static uint8_t  rx_line_count;
+static uint8_t rx_buf[RX_BUF_SIZE] __attribute__((aligned(256)));
+static uint8_t rx_head;
+static uint8_t rx_tail;
+static uint8_t rx_errs;
+static uint8_t rx_line_count;
 
 void init_serial(void)
 {
